backend/requirements.txt
fastapi
multipart
uvicorn
redis-om
google-generativeai
python-dotenv

python-multipart
python-jose
aiofiles
requests
sqlalchemy

